Make package.el complain less
If a (package-initialize) line isn't the first thing in init.el,
package.el will put it in when it first runs. It doesn't mind if it's
commented out though. ¯\_(ツ)_/¯

I don't impose package.el on anyone in the framework, and I know how
to enable it in my specific config code, so just make package.el happy
and move on.